+++ packages/fakta-inntektsmelding/.storybook/main.ts
@@ -1,0 +1,1 @@
+export { default } from '../../../.storybook/main-storybook';
