+++ packages/fakta-inntektsmelding/.storybook/preview.tsx
@@ -1,0 +1,1 @@
+export { decorators } from '../../../.storybook/preview-storybook';
