+++ packages/fakta-inntektsmelding/i18n/nb_NO.json
@@ -1,0 +1,3 @@
+{
+
+}
