+++ packages/fakta-inntektsmelding/index.ts
@@ -1,0 +1,1 @@
+export { InntektsmeldingFaktaIndex } from './src/InntektsmeldingFaktaIndex';
