+++ packages/fakta-inntektsmelding/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "@navikt/fp-fakta-inntektsmelding",
+  "license": "MIT",
+  "private": true,
+  "type": "module",
+  "main": "index.ts",
+  "types": "index.ts",
+  "scripts": {
+    "test": "vitest",
+    "test:watch": "vitest --watch=true",
+    "test:coverage": "vitest --coverage",
+    "tsc": "tsc --pretty",
+    "eslint": "eslint \"src/**/*.ts*\" --color",
+    "eslint:fix": "eslint --fix \"src/**/*.ts*\" --color",
+    "stylelint": "stylelint \"src/**/*.module.css\"",
+    "prettier": "prettier --write src",
+    "clean": "rm -rf ./dist ./node_modules ./coverage ./.storybook-static-build",
+    "build-storybook": "storybook build -o .storybook-static-build"
+  },
+  "dependencies": {
+    "@navikt/aksel-icons": "6.13.0",
+    "@navikt/ds-css": "6.13.0",
+    "@navikt/ds-react": "6.13.0",
+    "@navikt/fp-fakta-felles": "workspace:*",
+    "@navikt/fp-kodeverk": "workspace:*",
+    "@navikt/fp-types": "workspace:*",
+    "@navikt/ft-kodeverk": "2.4.14",
+    "@navikt/ft-types": "2.4.14",
+    "@navikt/ft-ui-komponenter": "2.3.17",
+    "@navikt/ft-utils": "2.3.16",
+    "react": "18.3.1",
+    "react-hook-form": "7.52.1",
+    "react-intl": "6.6.8"
+  },
+  "devDependencies": {
+    "@navikt/fp-storybook-utils": "workspace:*",
+    "@storybook/addon-actions": "8.2.7",
+    "@storybook/react": "8.2.7",
+    "@storybook/react-vite": "8.2.7",
+    "@testing-library/dom": "10.4.0",
+    "@testing-library/react": "16.0.0",
+    "@testing-library/user-event": "14.5.2",
+    "cross-env": "7.0.3",
+    "eslint": "9.8.0",
+    "storybook": "8.2.7",
+    "stylelint": "16.8.1",
+    "typescript": "5.5.4",
+    "vite": "5.3.5",
+    "vitest": "2.0.5"
+  }
+}
